+++ crates/rome_lsp/src/session.rs
@@ -381,7 +381,7 @@ impl Session {
         };
 
         let status = match load_config(&self.fs, base_path) {
-            Ok(Some(deserialized)) => {
+            Ok(Some((deserialized, _))) => {
                 let (configuration, diagnostics) = deserialized.consume();
                 if diagnostics.is_empty() {
                     warn!("The deserialization of the configuration resulted in errors. Rome will its defaults where possible.");