+++ crates/rome_migrate/src/registry.rs
@@ -1,0 +1,7 @@
+use crate::analyzers::MigrationCategory;
+use rome_analyze::RegistryVisitor;
+use rome_json_syntax::JsonLanguage;
+
+pub fn visit_migration_registry<V: RegistryVisitor<JsonLanguage>>(registry: &mut V) {
+    registry.record_category::<MigrationCategory>();
+}